fix: Prevent NPE on app resume
When the app resumes because Android temporarily killed it (i.e. due to
low memory) the class variable requestResults is null and trying to put
a result in it causes a NPE. In this scenario the Flutter app has been
restarted and will no longer be waiting for permissions so we can just
add a null check and leave it at that because no result will be returned
and Flutter is not waiting for one anyway.